auth_service/requirements.txt
bcrypt==4.2.1
Flask==3.1.0
mysql-connector-python==9.1.0
PyJWT==2.10.1